refactor: decouple package manager implementation
This decouples the package manager implementation from the public API.  The result is something akin to a "plugin system."

Right now, there's only the "npm" implementation, which is used by default.  This will be further built upon to support other package managers.

The `Smoker` class gets an object implementing the `PackageManager` interface, and that's about all it knows.

Also: more tests!  Replaced `unexpected-snapshot` with `snap-shot-it`, because the former does not work with TypeScript.