fix!: pin corepack version
BREAKING CHANGE: This package now requires a minimum Node.js of v16.20.0.

`corepack` v0.17.0 which ships with Node.js v16 seems completely broken (at least on my machine!) so we can't rely on the version shipped with Node.js.  This adds `corepack` as a dev dependency, and since the minimum version is Node.js v16.20.0, we have to bump our minimum version as well.

This also simplifies finding the `corepack` executable.