chore: setup starlight-typedoc
This is currently broken due to https://github.com/HiDeoo/starlight-typedoc/issues/23

To run example:

1. Clone repo
2. Checkout this branch (`boneskull/starlight-typedoc`)
3. Run `npm install`
4. Run `npm run dev:docs`

This should fail without any obvious error; this is because of the check for `groups` in `starlight-typedoc`.

You can also run `npm exec typedoc` from the workspace root to see that the problem is unique to `starlight-typedoc`.